requirements.txt
python-dotenv>=1.0.0
cassandra-driver>=3.27.0
fastapi>=0.95.0
flask>=2.2.0
openai>=1.0.0
pandas>=2.0.0
httpx>=0.24.0
uvicorn>=0.20.0
scipy>=1.10.0
scikit-learn>=1.2.0
